% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_auth.R
\docType{data}
\name{FacebookAuth}
\alias{FacebookAuth}
\title{R6 environment to store authentication credentials}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FacebookAuth
}
\description{
Used to keep persistent state.
}
\keyword{datasets}
