% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_adAccounts.R
\name{fb_adAccounts}
\alias{fb_adAccounts}
\title{List all ad accounts associated with the authorized user}
\usage{
fb_adAccounts()
}
\description{
List all ad accounts associated with the authorized user
}
\examples{
fb_adAccounts()
}
