% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_campaigns.R
\name{fb_campaigns}
\alias{fb_campaigns}
\title{List all campaigns for the chosen ad accounts}
\usage{
fb_campaigns(account_id = fb_adAccounts())
}
\arguments{
\item{account_id}{The id of one or more ad account ids (act_XXXXXXX)}
}
\description{
List all campaigns for the chosen ad accounts
}
\examples{
fb_campaigns(account_id = fb_adAccounts())
}
