% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_userId.R
\name{fb_userId}
\alias{fb_userId}
\title{Get the User ID for the authorized Facebook user}
\usage{
fb_userId()
}
\description{
This function will import the User ID for the authenticated user.
}
\examples{
fb_userId()
}
