% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_fb_data.R
\name{fetch_fb_data}
\alias{fetch_fb_data}
\title{Collect data from Facebook API}
\usage{
fetch_fb_data(request_string, api_version = "3.2",
  print.status = FALSE)
}
\arguments{
\item{request_string}{A string that contains the API call to GET from Facebook.}

\item{api_version}{The version of Facebook API.}
}
\description{
This function will import the data for the chosen parameters.
}
\examples{
fetch_fb_data(request_string, api_version="3.2")
}
