% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_stats.R
\name{fb_stats}
\alias{fb_stats}
\title{Import statistics from Facebook}
\usage{
fb_stats(account_id, fromDate, toDate, dimensions, metrics,
  level = "campaign")
}
\arguments{
\item{account_id}{The account id for which you want to import data fram (needs to start with "act_").}

\item{fromDate}{The starting date for the time range you want.}

\item{toDate}{The ending date for the time range you want.}

\item{dimensions}{The dimensions to use in the API request.}

\item{metrics}{The metrics to use in the API request.}

\item{level}{The level in which you want the data returned (defaults to "campaign")}
}
\description{
This function will import the statistics for the chosen parameters.
}
\examples{
fb_stats(account_id, fromDate, toDate, dimensions, metrics, level=c("account","campaign","adset","ad"))
}
