% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_fb_data.R
\name{fetch_fb_data}
\alias{fetch_fb_data}
\title{Collect data from Facebook API}
\usage{
fetch_fb_data(object, scope = "", parameters = "",
  api_version = "3.2")
}
\arguments{
\item{object}{The object to select data from.}

\item{scope}{The scope to select data from.}

\item{parameters}{The parameters to download. Breakdowns, dimensions and metrics.}

\item{api_version}{The version of Facebook API.}
}
\description{
This function will import the data for the chosen parameters.
}
\examples{
fetch_fb_data(object=c("me","accounts","campaigns",fb_userId()), scope=c("adaccounts","insights"), parameters="", api_version="3.2")
}
